Logfile now contains threadid for each entry. Dealt with several (non-critical) warnings provided by MS Code Analysis.

+++ OpenEBTS/OpenEBTS/NISTField.cpp
@@ -184,7 +184,7 @@ int CNISTField::RemoveSubField(int nSubField)
 
 		if (pItem->m_nSubField == nSubField)
 		{
-			wSubFieldIndex.push_back(i);
+			wSubFieldIndex.push_back((WORD)i);
 			bFound = true;
 		}
 	}
